# specs/001-build-a-comprehensive/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Amharic Document Preparation System API
  description: |
    Comprehensive API for processing, analyzing, and managing Amharic documents with multi-agent orchestration.
    
    ## Features
    - Multi-format document upload (PDF, images, Word, CSV, web content)
    - Real-time processing with CrewAI agents
    - Amharic-specific text analysis and OCR
    - Advanced search and export capabilities
    - Quality assurance and validation metrics
    
  version: 1.0.0
  contact:
    name: Amharic Document System Team
    email: support@amharic-docs.ai
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.amharic-docs.ai/v1
    description: Production server
...
paths:
...
  /mcp/tools:
    get:
      tags: [MCP]
      summary: List available MCP tools
      description: Get list of tools available for CopilotKit integration
      security:
        - BearerAuth: []
      responses:
        '200':
          description: MCP tools retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MCPTool'
  /mcp/tools/{tool_name}/execute:
    post:
      tags: [MCP]
      summary: Execute MCP tool
      security:
        - BearerAuth: []
      parameters:
        - name: tool_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPExecuteRequest'
      responses:
        '200':
          description: Tool executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPExecuteResponse'

  # WebSocket endpoint for real-time updates
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MCPTool:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        parameters:
          type: object
        returns:
          type: object
    MCPExecuteRequest:
      type: object
      properties:
        parameters:
          type: object
        context:
          type: object
    MCPExecuteResponse:
      type: object
      properties:
        result:
          type: object
        status:
          type: string
          enum: [success, error]
        message:
          type: string

    # Upload Response
...